LIBbb10309 If setRemindInterval is set to 25 or more, the number of days is not set correctly

+++ library/src/main/java/com/vorlonsoft/android/rate/AppRate.java
@@ -65,7 +65,7 @@ public final class AppRate {
     }
 
     private static boolean isOverDate(long targetDate, byte threshold) {
-        return new Date().getTime() - targetDate >= threshold * 24 * 60 * 60 * 1000;
+        return new Date().getTime() - targetDate >= (long) (threshold) * 24 * 60 * 60 * 1000;
     }
 
     public AppRate setLaunchTimes(@SuppressWarnings("SameParameterValue") byte launchTimes) {
